# swagger-docs/openapi.yaml
openapi: 3.0.0
info:
  title: MobyPark API
  version: 1.0.0
servers:
  - url: http://localhost:8000
paths:
...
  /vehicles:
    get:
      summary: List vehicles
      parameters:
        - name: Authorization
          in: header
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Vehicles list
    post:
      summary: Add vehicle
      parameters:
        - name: Authorization
          in: header
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                license_plate: { type: string }
              required: [name, license_plate]
      responses:
        '201':
          description: Vehicle added
  /vehicles/{id}:
    get:
      summary: Get vehicle
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
        - name: Authorization
          in: header
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Vehicle details
    put:
      summary: Update vehicle
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
        - name: Authorization
          in: header
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
      responses:
        '200':
          description: Vehicle updated
    delete:
      summary: Delete vehicle
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
        - name: Authorization
          in: header
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Vehicle deleted
...